slightly update wiktionary_en; new name for conceptnet5.stem
wiktionary_en.py has been ported to use the new organization of the
package. However, it has not been cleaned up or documented, because we
plan to replace it entirely.

'conceptnet5.stemmers' has been renamed to 'conceptnet5.stem', so it's
talking about the operation and not about whatever objects or functions
might perform it.