refactor the Wiktionary extraction step, and...
add a file for the parsing step. It doesn't do anything yet, it just
contains the functions that were previously in the extraction step but
don't belong there.

The extraction step no longer writes millions of files into your
readers/ directory. It instead takes a directory name on the command
line, like other readers do, and outputs a sensible number of .jsons
files, like other readers do.

The contents of the JSON structures are not edges; they're the
semi-structured content of Wiktionary definitions, in need of further
parsing.