Remove dependency on metanl.
metanl was created so it could support both ConceptNet and our
development at Luminoso. The use cases have diverged significantly,
however, and it no longer makes sense to support them with a single
package.

ConceptNet used only a small part of metanl, so we can instead
merge it in as the "conceptnet5.language" subpackage.