Disambiguate what language links are in
When a link points to a word in an unspecified language, we check
whether we have seen an entry for that word in either the language of
the headword being defined, or the language of the Wiktionary. If we
have seen one, we use it; if we see neither, we skip the link.