include a .tar.gz of APSW, making setup easier
APSW is an optional library for taking advantage of the extra features of
SQLite. It has to be installed separately, with its own process that
doesn't cooperate with pip.

Previously, the lib/ directory used a Makefile to download, unzip, and install
APSW. Now, the download step is unnecessary; we include a .tar.gz of APSW
(and therefore an unzip tool doesn't need to be installed, either).