rework 'valid concept names', handle exceptions smoothly
I've found that the existing normalize_text code is difficult to use
because it may raise a ValueError when given a text such as ','. It's
preferable to replace these strings with '_' so they can still be
represented in a URI, but still allow `valid_concept_name` to check
whether this would happen.