py2 compatibility; bring wordfreq's tokenizer into ConceptNet
This allows us to use wordfreq's simple_tokenize function without having
to depend on all of wordfreq. It replaces the NLTK 'punkt' tokenizer.
A slightly degraded version that works on Python 2 is also available.

In general, these changes enable ConceptNet 5.4 to work on Python 2 for
all steps except for building.