run eval on the downloaded vectors; expand range of analogy weights
On precomputed/numberbatch.h5, I've found that the analogy weights
of 0.2 and 0.6 fit the development set of analogies better than weights
I tried before. Because my methodology is to maximize the performance of
these weights on the development set, I let it use these weights to be
sure I'm following the methodology correctly.

In fact, these weights get one more question wrong on the complete
evaluation than the weights I would have preferred, but that's not my
decision to make.