Refactor SQL queries so they don't give up before matching things
We used to be pruning prefix matches to the first 200 matching nodes.
This totally doesn't work if you're trying to filter for nodes in a
particular language, as there are clearly going to be more than 200
nodes matching that prefix.

If we simplify the query, we don't need to do this pruning. The limit
at the end of 10000 results is enough to convince PostgreSQL's query
planner to do something reasonable, it seems, without having to
throw out valid matches.

A few queries take many seconds, such as {'node': '/c/en', 'other':
'/c/fr'}. But the API responses for these queries will be cached,
and then we're okay.