Log errors to Sentry when a file is present
This lets me copy the Sentry key as a file between ConceptNet servers. I
put it in the $CONCEPTNET_DATA directory because that's a place whose
path we already have to know, and so that I can just copy the Sentry key
when copying data.

Also added a mechanism for getting an error, to test that this works,
and made errors go to Sentry even in debug mode, so we can test errors
while developing.