Replace .loc[] with .reindex() when keys might be missing
There are many instances of .loc[] with a list that I have left as-is;
those are the ones where I think we *should* get an error if keys are
missing. But in the cases where we're just going to drop the missing
rows with .dropna(), we should be using .reindex().