We can handle more prefixes; don't strip off word sense info
We used to truncate concepts to 4 URI components, effectively removing
all the word sense information beyond the part of speech, as a way to
make prefix lookups efficient enough.

In addition to throwing out information that could be used for machine
learning, this was also throwing out sources of edges, in the case where
they created a duplicate URI because they produced the same assertion as
another edge but with different word senses.

Now that we have GIN, we probably don't need to do that, and we can
retain this information.